requirements.txt
Pillow>=10.0.0
imagehash>=4.3.1
piexif>=1.1.3
rawpy>=0.18.0  # optional, for RAW file support (CR2, NEF, ARW, DNG, etc.)
